Job Search Project/requirements.txt
chromadb
sentence-transformers
openai
streamlit
requests
python-dotenv
pypdf
pytest
